+++ packages/expo-sqlite/android/src/main/java/expo/modules/sqlite/SQLiteOptions.kt
@@ -16,5 +16,17 @@ internal data class OpenDatabaseOptions(
   val useNewConnection: Boolean = false,
 
   @Field
-  val finalizeUnusedStatementsBeforeClosing: Boolean = true
+  val finalizeUnusedStatementsBeforeClosing: Boolean = true,
+
+  @Field
+  val libSQLUrl: String? = null,
+
+  @Field
+  val libSQLAuthToken: String? = null,
+
+  @Field
+  val libSQLRemoteOnly: Boolean = false,
+
+  @Field
+  val libSQLSyncInterval: Int = 0
 ) : Record